# Performance: Transform `Utils::STATUS_WITH_NO_ENTITY_BODY` into a `Hash`