--- a/lib/rack/deflater.rb
+++ b/lib/rack/deflater.rb
@@ -107,7 +107,7 @@ module Rack
     def should_deflate?(env, status, headers, body)
       # Skip compressing empty entity body responses and responses with
       # no-transform set.
-      if Utils::STATUS_WITH_NO_ENTITY_BODY.include?(status.to_i) ||
+      if Utils::STATUS_WITH_NO_ENTITY_BODY.key?(status.to_i) ||
           headers['Cache-Control'].to_s =~ /\bno-transform\b/ ||
          (headers['Content-Encoding'] && headers['Content-Encoding'] !~ /\bidentity\b/)
         return false