## Change

`Utils::STATUS_WITH_NO_ENTITY_BODY` is a `Set` instance and Rack uses `#include?`. My assumption is that `Set` is used to guarantee uniqueness of values and to have faster access than `Array`.

This is a proposal to change that constant from a `Set` into a `Hash`. It guarantees uniqueness of keys, and slightly faster access via `#key?`.

Please note that the use of `true` is just a dummy value. But it's important to use a singleton method to keep low the memory footprint.

```ruby
#!/usr/bin/env ruby
# frozen_string_literal: true

require "benchmark/ips"
require "set"

STATUS_WITH_NO_ENTITY_BODY_SET  = Set.new((100..199).to_a << 204 << 304)
STATUS_WITH_NO_ENTITY_BODY_HASH = Hash[((100..199).to_a << 204 << 304).product([true])]

Benchmark.ips do |x|
  x.report("set")  { STATUS_WITH_NO_ENTITY_BODY_SET.include?(304) }
  x.report("hash") { STATUS_WITH_NO_ENTITY_BODY_HASH.key?(304) }
  x.compare!
end

__END__
Result:

  Warming up --------------------------------------
                 set   258.490k i/100ms
                hash   263.442k i/100ms
Calculating -------------------------------------
                 set      7.528M (± 4.9%) i/s -     37.740M in   5.025552s
                hash      8.345M (± 5.4%) i/s -     41.624M in   5.004100s

Comparison:
                hash:  8344683.5 i/s
                 set:  7528352.1 i/s - same-ish: difference falls within error

Ruby:

    ruby 2.5.3p105 (2018-10-18 revision 65156) [x86_64-darwin16]

Hardware:

    Hardware Overview:

      Model Name: MacBook Pro
      Model Identifier: MacBookPro12,1
      Processor Name: Intel Core i7
      Processor Speed: 3,1 GHz
      Number of Processors: 1
      Total Number of Cores: 2
      L2 Cache (per Core): 256 KB
      L3 Cache: 4 MB
      Memory: 16 GB
      Boot ROM Version: 180.0.0.0.0
      SMC Version (system): 2.28f7

Software:

    System Software Overview:

      System Version: macOS 10.12.6 (16G1618)
      Kernel Version: Darwin 16.7.0
      Time since boot: 17 days 28 minutes
```