@rafaelfranca 2019-01-03
Given the benchmark shows that the difference falls within error it is hard to judge if there is any performance difference for this change. Were you able to see this in different benchmarks?