@jodosha 2019-01-03
@rafaelfranca Not really, but I still proposed this change because the number of i/s is slightly better (+810K~) for `Hash` instead of `Set`.

I'll leave it to your judgment. Thanks for having a look at all of them. 👍