@rafaelfranca 2019-01-03
I got uncomfortable with the values being just a dummy value.

Looking the implementation of `Set#include?` and `Hash#key?` there is only the overhead of a `Hash#[]` method call in `Set#include?` when the element is on the set, given `Hash#[]` and `Hash#key?` are almost the same for when the element exists in the hash. For the worst case (when the element is not in the set) the `Hash#key?` would be faster because `Hash#[]` checks for the default value while `key?` just return false.

So for the case where the response has a method body it is indeed 20% faster:

```
#!/usr/bin/env ruby
# frozen_string_literal: true

require "benchmark/ips"
require "set"

STATUS_WITH_NO_ENTITY_BODY_SET  = Set.new((100..199).to_a << 204 << 304)
STATUS_WITH_NO_ENTITY_BODY_HASH = Hash[((100..199).to_a << 204 << 304).product([true])]

Benchmark.ips do |x|
  x.report("set")  { STATUS_WITH_NO_ENTITY_BODY_SET.include?(500) }
  x.report("hash") { STATUS_WITH_NO_ENTITY_BODY_HASH.key?(500) }
  x.compare!
end

__END__
Warming up --------------------------------------
                 set   323.563k i/100ms
                hash   363.396k i/100ms
Calculating -------------------------------------
                 set     12.212M (± 3.5%) i/s -     61.153M in   5.014672s
                hash     15.116M (± 4.6%) i/s -     75.586M in   5.012996s

Comparison:
                hash: 15116435.9 i/s
                 set: 12211765.4 i/s - 1.24x  slower
```